requirements.txt
asgiref==3.3.0
Django==3.0.5
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djongo==1.3.3
dnspython==2.0.0
PyJWT==1.7.1
pymongo==3.11.0
pytz==2020.1
sqlparse==0.2.4
whitenoise==5.2.0
